[PATCH 1/11] Add JSON validation to aws_s3_bucket.
This commit adds ValidateFunc to the policy attribute so that JSON parsing
errors can be caught early. Generally, when there is a ValidateFunc set for the
attribute, one can safely assume that before any of the creation and/or update
of the existing resource would happen it would have to succeed validation.

Resources affected (AWS):

  - aws_cloudformation_stack
  - aws_cloudwatch_event_rule
  - aws_elasticsearch_domain
  - aws_glacier_vault
  - aws_kms_key
  - aws_s3_bucket
  - aws_sns_topic
  - aws_sqs_queue
  - aws_vpc_endpoint
  - opsworks_layers

Resources affected (AzureRM):

  - arm_template_deployment

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>